Now Matrix<int>

+++ MATRIX/main_matrix.cpp
@@ -6,7 +6,7 @@ int main()
 {
 	int size = 0;
 	std::cin >> size;
-	matrix<long double> test(size);
+	matrix<int> test(size);
 	test.fill_M();
 	long double det = test.determinate();
 	std::cout << det << std::endl;
